+++ src/client/intentClient.ts
@@ -30,14 +30,17 @@ import type { GetUserIntentStatusResult } from "../actions/getUserIntentStatus.j
 import type { SendUserIntentResult } from "../actions/sendUserIntent.js";
 import type { GetUserIntentReceiptResult } from "../actions/types.js";
 import { ZERODEV_URLS } from "../config/constants.js";
-import { type CabClientActions, cabClientActions } from "./decorators/cab.js";
+import {
+  type IntentClientActions,
+  intentClientActions,
+} from "./decorators/intent.js";
 
 export type IntentRpcSchema = [
   {
     Method: "ui_getIntent";
     Parameters: [GetIntentParameters];
     ReturnType: GetIntentReturnType;
-  },
+  }
 ];
 
 // Relayer methods
@@ -61,34 +64,34 @@ export type RelayerRpcSchema = [
     Method: "rl_getUserIntentExecutionReceipt";
     Parameters: [Hex];
     ReturnType: GetUserIntentReceiptResult;
-  },
+  }
 ];
 
-// Combined schema for the CAB client
-export type CabRpcSchema = [...IntentRpcSchema, ...RelayerRpcSchema];
+// Combined schema for the Intent client
+export type CombinedIntentRpcSchema = [...IntentRpcSchema, ...RelayerRpcSchema];
 
-export type CabClient<
+export type IntentClient<
   transport extends Transport = Transport,
   chain extends Chain | undefined = Chain | undefined,
   account extends SmartAccount | undefined = SmartAccount | undefined,
   client extends Client | undefined = Client | undefined,
-  rpcSchema extends RpcSchema | undefined = undefined,
+  rpcSchema extends RpcSchema | undefined = undefined
 > = Prettify<
   Client<
     transport,
     chain extends Chain
       ? chain
       : // biome-ignore lint/suspicious/noExplicitAny: <explanation>
-        client extends Client<any, infer chain>
-        ? chain
-        : undefined,
+      client extends Client<any, infer chain>
+      ? chain
+      : undefined,
     account,
     rpcSchema extends RpcSchema
-      ? [...rpcSchema, ...CabRpcSchema]
-      : CabRpcSchema,
+      ? [...rpcSchema, ...CombinedIntentRpcSchema]
+      : CombinedIntentRpcSchema,
     BundlerActions<account> &
       KernelAccountClientActions<chain, account> &
-      CabClientActions<chain, account>
+      IntentClientActions<chain, account>
   >
 > & {
   client: client;
@@ -97,44 +100,46 @@ export type CabClient<
   userOperation: BundlerClientConfig["userOperation"] | undefined;
 };
 
-export type CreateCabClientConfig<
+export type CreateIntentClientConfig<
   transport extends Transport = Transport,
   chain extends Chain | undefined = Chain | undefined,
   account extends SmartAccount | undefined = SmartAccount | undefined,
   client extends Client | undefined = Client | undefined,
-  rpcSchema extends RpcSchema | undefined = undefined,
+  rpcSchema extends RpcSchema | undefined = undefined
 > = SmartAccountClientConfig<transport, chain, account, client, rpcSchema> & {
   bundlerTransport: transport;
   intentTransport?: transport;
   relayerTransport?: transport;
 };
 
-export function createCabClient<
+export function createIntentClient<
   transport extends Transport,
   chain extends Chain | undefined = undefined,
   account extends SmartAccount | undefined = undefined,
   client extends Client | undefined = undefined,
-  rpcSchema extends RpcSchema | undefined = undefined,
+  rpcSchema extends RpcSchema | undefined = undefined
 >(
-  parameters: CreateCabClientConfig<
+  parameters: CreateIntentClientConfig<
     transport,
     chain,
     account,
     client,
     rpcSchema
-  >,
-): CabClient<transport, chain, account, client, rpcSchema>;
+  >
+): IntentClient<transport, chain, account, client, rpcSchema>;
 
-export function createCabClient(parameters: CreateCabClientConfig): CabClient {
+export function createIntentClient(
+  parameters: CreateIntentClientConfig
+): IntentClient {
   const {
     client: client_,
     key = "Account",
-    name = "CAB Client",
+    name = "Intent Client",
     paymaster,
     paymasterContext,
     bundlerTransport,
     intentTransport = http(ZERODEV_URLS.INTENT_SERVICE),
-    relayerTransport = http(ZERODEV_URLS.RELAYER_SERVICE),
+    relayerTransport = http(ZERODEV_URLS.RELAYER_SERVICE_MAINNET),
     userOperation,
   } = parameters;
 
@@ -168,9 +173,9 @@ export function createCabClient(parameters: CreateCabClientConfig): CabClient {
       transport,
       key,
       name,
-      type: "cabClient",
+      type: "intentClient",
     }),
-    { client: client_, paymaster, paymasterContext, userOperation },
+    { client: client_, paymaster, paymasterContext, userOperation }
   );
 
   if (parameters.userOperation?.prepareUserOperation) {
@@ -184,11 +189,11 @@ export function createCabClient(parameters: CreateCabClientConfig): CabClient {
           return customPrepareUserOp(client, args);
         },
       }))
-      .extend(cabClientActions()) as CabClient;
+      .extend(intentClientActions()) as IntentClient;
   }
 
   return client
     .extend(bundlerActions)
     .extend(kernelAccountClientActions())
-    .extend(cabClientActions()) as CabClient;
+    .extend(intentClientActions()) as IntentClient;
 }