+++ src/config/constants.ts
@@ -2,7 +2,8 @@ import type { Address } from "viem";
 
 export const ZERODEV_URLS = {
   INTENT_SERVICE: "https://user-intent-service.onrender.com/intent",
-  RELAYER_SERVICE: "https://relayer-d6ne.onrender.com",
+  RELAYER_SERVICE_MAINNET: "https://relayer-d6ne.onrender.com",
+  RELAYER_SERVICE_TESTNET: "https://relayer-testnet.onrender.com",
 } as const;
 
 export const INTENT_EXECUTOR: Address =