+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-# @zerodev/cab-sdk
+# @zerodev/intent
 
 ## 0.0.1
 
 ### Patch Changes
 
-- feat: get and send intent
+- Initial release